Move further files to km; split off preload / ipc to km
